fix(k8s): fixed the name of the example function for TestConnection
The `ExampleTestConnection` function was throwing a `refers to unknown identifier` error when running `go test`. This was due to the fact that its name did not match the Go naming convention for example functions that demonstrate a method of a certain type.

The function was renamed to `ExampleClient_TestConnection` to correctly associate it with the `TestConnection` method of the `Client` structure, which solves the test compilation error.